requirements.txt
Django==2.0.1
django-webpack-loader==0.5.0
djangorestframework==3.7.7
drf-role==0.1.0
Pillow==5.0.0
pytz==2017.3
psycopg2==2.7.3
